permissao

+++ app/Http/Controllers/UserController.php
@@ -6,6 +6,7 @@ use App\Models\Role;
 use App\Models\User;
 use App\Models\UserProfile;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class UserController extends Controller
 {
@@ -45,6 +46,7 @@ class UserController extends Controller
 
     public function edit(User $user)
     {
+        Gate::authorize('edit', User::class);
         $user->load(['profile', 'interests']);
         $roles = Role::all();
         return view('users.edit', compact('user', 'roles'));
@@ -52,6 +54,7 @@ class UserController extends Controller
 
     public function update(Request $request, User $user)
     {
+        Gate::authorize('edit', User::class);
         $input = $request->validate([
             'name' => 'required',
             'email' => 'required|email',
@@ -65,6 +68,7 @@ class UserController extends Controller
 
     public function updateProfile(Request $request, User $user)
     {
+        Gate::authorize('edit', User::class);
         $input = $request->validate([
             'type' => 'required',
             'address' => 'nullable',
@@ -81,6 +85,7 @@ class UserController extends Controller
     public function updateInterests(Request $request, User $user)
     {
 
+        Gate::authorize('edit', User::class);
         $input = $request->validate([
             'interests' => 'nullable|array',
         ]);
@@ -108,6 +113,7 @@ class UserController extends Controller
 
     public function destroy(User $user)
     {
+        Gate::authorize('edit', User::class);
         $user->delete();
         return back()->with('success', 'Usuário deletado com sucesso.');
     }
